image_insertion/requirements.txt
boto3>=1.38.0
Pillow>=11.0.0
